Tags: Reduce subscriber field return
A subscriber being tagged can be done without a secret, this key is
semi-public so it's a security issue that we're returning all the
sensitive fields for a subscriber as the response of a tag request.

We're just removing the fields from the documentation for now since it's
possible for an API to get all those fields from the subscriber endpoint
which requires the secret key.